Chore/extract add tokens components & functions (#2465)
