+++ packages/checkout/widgets-lib/src/functions/getDurationFormatted.ts
