+++ packages/checkout/widgets-lib/src/types/HandoverTypes.ts
@@ -1,0 +1,7 @@
+export enum RiveStateMachineInput {
+  START = 0,
+  WAITING = 1,
+  PROCESSING = 2,
+  COMPLETED = 3,
+  ERROR = 4,
+}
