+++ packages/checkout/widgets-lib/src/widgets/add-tokens/types.ts
@@ -3,14 +3,6 @@ export enum FiatOptionType {
   DEBIT = 'debit',
 }
 
-export enum RiveStateMachineInput {
-  START = 0,
-  WAITING = 1,
-  PROCESSING = 2,
-  COMPLETED = 3,
-  ERROR = 4,
-}
-
 export type AddTokensError = {
   type: AddTokensErrorTypes;
   data?: Record<string, unknown>;
